Revert "(broken) implement instruction grouping"
This reverts commit 4b8d503d3e63b9e05258ccd52a356e3de1097739.
When grouping multiple actions together, desmos does not update
the values in sequence. For example, consider the following
situation:
a=1
b=2
(a->42,b->a)
One would expect that the value of b would be 42 after executing
this statement, but in reality it would be 1 for some reason.
For this reason, I have to revert all of my work on instruction
grouping. Oh well.