Revert "Implement reset function"
This reverts commit 3738f5013beab42186584d51ceb8804e76cbd34e.
Reset is impossible to implement because once a function doesn't
produce an action once, it is done for good and you can't run it
ever again, as far as I know. Even if you put a junk action to
keep the timer going, the update function will just never be run
again, ever after resetting.

In conclusion, the only way to truly reset it to load the calculator
state again, whether that be by refreshing or whatever.